requirements.txt
fastapi 
uvicorn 
sentence-transformers
faiss-cpu
whoosh 
exa_py
python-dotenv